% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_add_word.R
\name{mod_add_word_ui}
\alias{mod_add_word_ui}
\alias{mod_add_word_server}
\title{mod_add_word_ui and mod_add_word_server}
\usage{
mod_add_word_ui(id)

mod_add_word_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
