% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc_word_vector.R
\name{mod_calc_word_vector_ui}
\alias{mod_calc_word_vector_ui}
\alias{mod_calc_word_vector_server}
\title{mod_calc_word_vector_ui and mod_calc_word_vector_server}
\usage{
mod_calc_word_vector_ui(id)

mod_calc_word_vector_server(input, output, session, react_add, react_subtract)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
