% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nearest_words_table.R
\name{mod_nearest_words_table_ui}
\alias{mod_nearest_words_table_ui}
\alias{mod_nearest_words_table_server}
\title{mod_nearest_words_table_ui and mod_nearest_words_table_server}
\usage{
mod_nearest_words_table_ui(id)

mod_nearest_words_table_server(input, output, session, react_wv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
