% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_subtract_word.R
\name{mod_subtract_word_ui}
\alias{mod_subtract_word_ui}
\alias{mod_subtract_word_server}
\title{mod_subtract_word_ui and mod_subtract_word_server}
\usage{
mod_subtract_word_ui(id)

mod_subtract_word_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
